+++ packages/plugin-packages/mashroom-portal/src/backend/controllers/PortalAppController.ts
@@ -212,7 +212,7 @@ export default class PortalAppController {
         const cacheControlService: MashroomCacheControlService = req.pluginContext.services.browserCache && req.pluginContext.services.browserCache.cacheControl;
 
         if (cacheControlService) {
-            await cacheControlService.addCacheControlHeader(false, req, res);
+            cacheControlService.addCacheControlHeader('ALWAYS', req, res);
         }
 
         const resourceUri = `${portalApp.resourcesRootUri}/${resourcePath}`;